新增UT编译选项，CM心跳时间修正


Match-id-997a672d34bc315afbd404207f953e6f72732491

+++ src/cluster/common/cm_inner.h
@@ -249,8 +249,8 @@ typedef struct {
 typedef struct {
     char *zkIpMask;
     char *ipStr;
-    uint16_t regTimeOut;
-    uint16_t regPermTimeOut;
+    uint32_t regTimeOut; // 毫秒
+    uint32_t regPermTimeOut; // 毫秒
 } CmCfgInfo;
 
 /*
